requirements.txt
prompt-toolkit==3.0.2
python-chess==0.30.0
wcwidth==0.1.8
